move artifactory config to options bundle to conform to architecture
